Exception No Python Macro removed (22 macros left)
