+++ torch/csrc/Exceptions.h
@@ -4,8 +4,6 @@
 #include <stdexcept>
 #include <string>
 
-#ifndef NO_PYTHON
-
 #include "ATen/Error.h"
 #include "THP_export.h"
 #include "torch/csrc/utils/object_ptr.h"
@@ -135,22 +133,3 @@ struct ValueError : public PyTorchError {
 };
 
 } // namespace torch
-
-#else
-
-namespace torch {
-
-struct PyTorchError : public std::exception {
-  virtual const char* what() const noexcept override {
-    return msg.c_str();
-  }
-  std::string msg;
-};
-
-struct ValueError : public PyTorchError {
-  ValueError(const char *format, ...);
-};
-
-} // namespace torch
-
-#endif
